requirements.txt
beautifulsoup4
discord.py
lxml
python-dotenv
requests